Move box/violin/stripplot code to new categorical module
