+++ seaborn/__init__.py
@@ -2,6 +2,7 @@ from .rcmod import *
 from .utils import *
 from .palettes import *
 from .linearmodels import *
+from .categorical import *
 from .distributions import *
 from .timeseries import *
 from .matrix import *
